third_party/blink/web_tests/external/wpt/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent-expected.txt

This is a testharness.js-based test.
[FAIL] Fulfillment handler on fulfilled promise, using backup incumbent settings object stack
  assert_equals: expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/resources/promise-job-incumbent-incumbent.html" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent.html"
[FAIL] Rejection handler on rejected promise, using backup incumbent settings object stack
  assert_equals: expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/resources/promise-job-incumbent-incumbent.html" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent.html"
[FAIL] Fulfillment handler on pending-then-fulfilled promise, using backup incumbent settings object stack
  assert_equals: expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/resources/promise-job-incumbent-incumbent.html" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent.html"
[FAIL] Rejection handler on pending-then-rejected promise, using backup incumbent settings object stack
  assert_equals: expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/resources/promise-job-incumbent-incumbent.html" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent.html"
[FAIL] Thenable resolution, using backup incumbent settings object stack
  assert_equals: expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/resources/promise-job-incumbent-incumbent.html" but got "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/integration-with-the-javascript-job-queue/promise-job-incumbent.html"
Harness: the test ran to completion.
